+++ build.gradle
@@ -12,7 +12,7 @@ apply plugin: 'net.minecraftforge.gradle'
 apply plugin: 'eclipse'
 
 version = "1.3-alpha"
-group = "com.milamberBrass.brass_armory"
+group = "com.milamber_brass.brass_armory"
 archivesBaseName = "brass_armory-1.16.5"
 
 java.toolchain.languageVersion = JavaLanguageVersion.of(8)
