+++ src/main/java/com/milamber_brass/brass_armory/blocks/RopeBlock.java
@@ -1,4 +1,4 @@
-package com.milamberBrass.brass_armory.blocks;
+package com.milamber_brass.brass_armory.blocks;
 
 import net.minecraft.block.*;
 import net.minecraft.fluid.FluidState;
