+++ src/main/java/com/milamber_brass/brass_armory/entities/render/BAArrowRenderer.java
@@ -1,8 +1,8 @@
-package com.milamberBrass.brass_armory.entities.render;
+package com.milamber_brass.brass_armory.entities.render;
 
-import com.milamberBrass.brass_armory.BrassArmory;
-import com.milamberBrass.brass_armory.entities.BAArrowEntity;
-import com.milamberBrass.brass_armory.util.ArrowType;
+import com.milamber_brass.brass_armory.BrassArmory;
+import com.milamber_brass.brass_armory.entities.BAArrowEntity;
+import com.milamber_brass.brass_armory.util.ArrowType;
 import net.minecraft.client.renderer.entity.ArrowRenderer;
 import net.minecraft.client.renderer.entity.EntityRendererManager;
 import net.minecraft.util.ResourceLocation;
