+++ src/main/java/com/milamber_brass/brass_armory/entities/render/SpearEntityRenderer.java
@@ -1,8 +1,8 @@
-package com.milamberBrass.brass_armory.entities.render;
+package com.milamber_brass.brass_armory.entities.render;
 
-import com.milamberBrass.brass_armory.BrassArmory;
-import com.milamberBrass.brass_armory.entities.SpearEntity;
-import com.milamberBrass.brass_armory.entities.model.SpearModel;
+import com.milamber_brass.brass_armory.BrassArmory;
+import com.milamber_brass.brass_armory.entities.SpearEntity;
+import com.milamber_brass.brass_armory.entities.model.SpearModel;
 import com.mojang.blaze3d.matrix.MatrixStack;
 import com.mojang.blaze3d.vertex.IVertexBuilder;
 import net.minecraft.client.renderer.IRenderTypeBuffer;
