+++ src/main/java/com/milamber_brass/brass_armory/items/SpearItem.java
@@ -1,10 +1,10 @@
-package com.milamberBrass.brass_armory.items;
+package com.milamber_brass.brass_armory.items;
 
 
 import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.Multimap;
-import com.milamberBrass.brass_armory.BrassArmory;
-import com.milamberBrass.brass_armory.entities.SpearEntity;
+import com.milamber_brass.brass_armory.BrassArmory;
+import com.milamber_brass.brass_armory.entities.SpearEntity;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.material.Material;
 import net.minecraft.client.Minecraft;
